requirements.txt
streamlit
pandas
plotly.express
base64
os